Hotfix: 컨트롤러 클래스 기능 분리
