--- src/main/java/com/nhnacademy/marketgg/auth/controller/TokenController.java
@@ -1,4 +1,0 @@
-package com.nhnacademy.marketgg.auth.controller;
-
-public class TokenController {
-}
