+++ src/main/java/com/nhnacademy/marketgg/auth/util/MailUtils.java
@@ -1,9 +1,6 @@
 package com.nhnacademy.marketgg.auth.util;
 
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
+import java.util.Properties;
 import javax.mail.Authenticator;
 import javax.mail.Message;
 import javax.mail.PasswordAuthentication;
@@ -11,7 +8,9 @@ import javax.mail.Session;
 import javax.mail.Transport;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
-import java.util.Properties;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 @Slf4j
 @Component
@@ -59,12 +58,13 @@ public class MailUtils {
             message.setSubject("[Market GG] 인증코드 전송");
 
             message.setContent(
-                    "<h1>[이메일 인증]</h1> <p>아래 버튼을 클릭하시면 이메일 인증이 완료됩니다.</p> " +
-                            "<form action=\"https://\"" + toHost + ":\"" + toPort + "/auth/use/email/use?use=true\" method=\"post\">\n" +
-                            "<input type=\"hidden\""+ "id=\"email\" value=\""+email+"\">" +
-                            "    <button type=\"submit\">인증하기</button>\n" +
-                            "</form>"
-                    , "text/html;charset=euc-kr"
+                "<h1>[이메일 인증]</h1> <p>아래 버튼을 클릭하시면 이메일 인증이 완료됩니다.</p> " +
+                    "<form action=\"https://\"" + toHost + ":\"" + toPort +
+                    "/auth/use/email/use?use=true\" method=\"post\">\n" +
+                    "<input type=\"hidden\"" + "id=\"email\" value=\"" + email + "\">" +
+                    "    <button type=\"submit\">인증하기</button>\n" +
+                    "</form>"
+                , "text/html;charset=euc-kr"
             );
 
             Transport.send(message);
